Decode numberic character references (decimal/hexadecimal)
I've changed the regex to capture the codepoint inside (without the '&''
prefix and ';' suffix). Because of that I had to change the arity of the
replace_entity function to 2. The original two declarations ignore the
second argument.

I added two replace_entity declarations. The first handles the
hexadecimal numberic character references, the second the decimal ones.
The way the hexadecimal value is turned into an integer should be
maybe changed to something more elegant.