Streamline decoding and encoding
This PR replaces the regex and graphemes use in favor
of binary matching. This makes decoding 10x faster and
encoding around 7x faster.

The PR also makes it so we don't load all entities into
memory when parsing the file but this is mostly likely
irrelevant since the file is quite small.